argos/argos_node/requirements.txt
# pyrealsense2==2.55.1.6486
Flask==3.0.3
flask_socketio==5.3.6
jsonschema==4.23.0
numpy==2.0.1
python-dotenv==1.0.1
PyYAML==6.0.1
appdirs==1.4.4